internal/serverstate: move job tests
